Split V4 release notes into separate page
The page being called "Release Notes V3" in the table of contents made
it difficult for me to find the V4 release notes for a while.

This commit also explicitly mentions the version number in each release
notes file name, to make it more difficult to forget to create a new
file for the next major version's release notes (the previous V3 release
notes file was just named `Adapter-Release-Notes.md`).